Split distributed OOMptimizer script
